Try to move code to separate class.
